reduce number of started goroutines
While the impact of starting goroutines is very low compared to regular
threads, it's still significant compared to sending a value on a
channel. This reduces ressource consumption drastically while keeping
runtimes almost identical.

Replace the emphemeral parseMessage function with a long-lived
messageParser function which receives new jobs from a channel.

Remove semaphore and move the closing of channels to the sender-side.

Remove tmp from scanned directories as messages in there have not been
transferred completely.

Signed-off-by: Moritz Poldrack <git@moritz.sh>