make output deterministic
There are two sources of indeterminism in the output: a) sorting by
ranks when ranks are equal and b) getting the most frequent name when
two names have the highest count. Implement sorting alphabetically based
on email address when ranks are equal (for all three types of ranks).
Implement sorting alphabetically when the names have the same count.

Signed-off-by: Bence Ferdinandy <bence@ferdinandy.com>