Prioritize non-empty names
It's nice to have a display name with an email address, even if that
address has often been used without one. Don't include empty names
(empty string after trimming spaces) when deciding which name to use for
an address. Fall back to using the empty string if no other name has
been used with the address.