output: add NormalizedName
Since mra uses the most frequent name used by each correspondent and
people with non-ascii names sometimes use the non-ascii version of their
names when emailing and sometimes they convert to ascii it's hard to
know what exactly to search for. Included NormalizedName in AddressData
which is unicode normalized. If the software reading the address book
allows for extra information for an entry, this field can be used to
search for both the ascii and the proper unicode version of a name as
well.