chore: rename e2e tests explicitly
The idea is that this is not a unittest, but is basically running
everything (except for maybe configuration and actual printing). It will
likely turn out that all other tests will not even need testdata, but
let's leave it this way in case they do. Since any new email added to
the e2e testdata could potentially mess up _all_ the tests by changing
the rankings of an address it is best to keep it separated from trying
to check cornercases of tests.

It also unlikely we actually will want to make unittests reading from
email files: testing the capabilities of the go mail parsing stack is
not our job.