addressbook: move logic to the very end
In hindsight it was unfortunate that using the addressbook was inserted
into the middle of parsing the mail envelopes, leading to a bit of
a clunky logic with checking for an empty name. Move the entire
application of the addressbook override to the very end where we figure
out the most frequent name as well. This will also provide a clean place
to put any further overrides to the most frequent name.