output: special case addresses that match list-id
Some lists do weird things with name, like adding "X Y via list" to the
From name, which is in fact the address name. This usually means that
m-r-a will pick up the maintainer's name as the list name. Instead of
doing this, for lists, use the List-Id header which has information
about the list. Allow the user to configure this somewhat.