walkpaths: fix messageParser not closing files
Using defer to close the file in general was a smart move, but I did not
consider that all the parsing happens within the same function, thus
leaving all the defer statements to after all the files have been
opened. Create an emlParser in analogue to the mboxParser, opening each
file in an individual function call to make use of defer. Make the
messageParser logic a bit cleaner in the process.

Reported-by: Federico Ferri <federico.ferri@cern.ch>
Fixes: 7e0f2bb9b5b8ab94eecad51d581e7b9c041cdffc
Closes: #14